resolved bug with mv to flag directory

+++ src/main.py
@@ -8,8 +8,11 @@ from colorama import Fore, Style
 parser = argparse.ArgumentParser()
 parser.add_argument('-p', '--package', required=False, action='store')
 parser.add_argument('-d', '--download', required=False, action='store')
-parser.add_argument('-c', '--checker', required=False, action='store')
-args = parser.parse_args()
+parser.add_argument('-c', '--checker', required=False, action='store', help='The Python file should be present in the "checker" directory.')
+try:
+    args = parser.parse_args()
+except argparse.ArgumentError as e:
+    print(e)
 
 path = os.getcwd()
 path = os.path.dirname(path)
@@ -55,21 +58,23 @@ elif args.checker is not None:
     path = path + '/checker'
     susp = False
    
-    if args.checker in os.listdir(path):
-        if os.path.isfile(path+'/'+args.checker) and pathlib.Path(path+'/'+args.checker).suffix == '.py':
+    if args.checker in os.listdir(path) and os.path.isfile(path+'/'+args.checker) and pathlib.Path(path+'/'+args.checker).suffix == '.py':
             
-            clean = remove_comments(path, args.checker)
-            if find_hardcoded_urls(clean) and not url_in_setup(clean) and not url_in_prints(clean):
-                susp = True
-                print(f"{Fore.RED}This file is suspicious.{Style.RESET_ALL}")
-            if not susp and manual_pip_install(clean):
-                susp = True
-                print(f"{Fore.RED}This file is suspicious.{Style.RESET_ALL}")
+        clean = remove_comments(path, args.checker)
+
+        if find_hardcoded_urls(clean) and not url_in_setup(clean) and not url_in_prints(clean):
+            susp = True
+            print(f"{Fore.RED}This file is suspicious.{Style.RESET_ALL}")
+        if not susp and manual_pip_install(clean):
+            susp = True
+            print(f"{Fore.RED}This file is suspicious.{Style.RESET_ALL}")
             
-            if not susp:
-                print(f"{Fore.GREEN}This file is seems to be safe.{Style.RESET_ALL}")
+        if not susp:
+            print(f"{Fore.GREEN}This file is seems to be safe.{Style.RESET_ALL}")
             
-            os.remove(path + '/' + clean)
+        os.remove(path + '/' + clean)
+    else:
+        print('This file is not present in "checker" directory or is not a Python file.')
             
 else:
     #default
